Group all sequelize connection parameters into single parameter
